+++ src/Symfony/Component/HttpClient/Exception/EmptyResponseBodyException.php
@@ -11,13 +11,9 @@
 
 namespace Symfony\Component\HttpClient\Exception;
 
-use Symfony\Contracts\HttpClient\Exception\DecodingExceptionInterface;
-
 /**
- * Thrown by responses' toArray() method when their content cannot be JSON-decoded.
- *
- * @author Nicolas Grekas <p@tchwork.com>
+ * Thrown by responses' toArray() method when the response body is empty.
  */
-class JsonException extends \JsonException implements DecodingExceptionInterface
+final class EmptyResponseBodyException extends JsonException
 {
 }