+++ src/Symfony/Component/HttpClient/Exception/JsonNotArrayException.php
@@ -11,13 +11,9 @@
 
 namespace Symfony\Component\HttpClient\Exception;
 
-use Symfony\Contracts\HttpClient\Exception\DecodingExceptionInterface;
-
 /**
- * Thrown by responses' toArray() method when their content cannot be JSON-decoded.
- *
- * @author Nicolas Grekas <p@tchwork.com>
+ * Thrown by responses' toArray() method when the decoded JSON is not an array.
  */
-class JsonException extends \JsonException implements DecodingExceptionInterface
+final class JsonNotArrayException extends JsonException
 {
 }