+++ src/Symfony/Component/HttpClient/CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG
 8.1
 ---
 
+ * Add `EmptyResponseBodyException`, `JsonDecodeException` and `JsonNotArrayException`
+   specific subclasses of `JsonException`, thrown by `ResponseInterface::toArray()` to let
+   callers distinguish the three decoding failure modes by type
  * Add support for the `max_connect_duration` option
  * Add option `extra.use_persistent_connections` to `CurlHttpClient` to control the use of persistent connections introduced in PHP 8.5
  * Add `GuzzleHttpHandler` that allows using Symfony HttpClient as a Guzzle handler